tesla_shop/requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
boto3==1.34.157
botocore==1.34.157
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
dj-rest-auth==6.0.0
django-allauth==0.63.6
django-cors-headers==4.4.0
django-filter==24.2
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
ffmpeg==1.4
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
inflection==0.5.1
infobip-api-python-client==4.0.0
jax==0.4.26
jaxlib==0.4.26
Jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
kiwisolver==1.4.5
kombu==5.3.7
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib==3.8.4
mediapipe==0.10.11
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
openai==1.51.0
opencv-contrib-python==4.9.0.80
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
phonenumbers==8.13.38
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
sqlparse==0.5.1
stripe==10.7.0
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.40
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.0.7
utils==1.0.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
